Fix concurrency issues in tests
