+++ Papyrus/Sources/Provider.swift
@@ -84,48 +84,3 @@ public protocol Interceptor: Sendable {
 public protocol RequestModifier {
     func modify(req: inout RequestBuilder) throws
 }
-
-private extension Provider {
-    // Note: Can be replaced with Synchronization framework starting with iOS 18.
-    final class ResourceMutex<R>: @unchecked Sendable {
-        private var resource: R
-        private let mutex: UnsafeMutablePointer<pthread_mutex_t>
-
-        init(resource: R) {
-            let mutexAttr = UnsafeMutablePointer<pthread_mutexattr_t>.allocate(capacity: 1)
-            pthread_mutexattr_init(mutexAttr)
-            pthread_mutexattr_settype(mutexAttr, Int32(PTHREAD_MUTEX_RECURSIVE))
-            mutex = UnsafeMutablePointer<pthread_mutex_t>.allocate(capacity: 1)
-            pthread_mutex_init(mutex, mutexAttr)
-            pthread_mutexattr_destroy(mutexAttr)
-            mutexAttr.deallocate()
-            self.resource = resource
-        }
-
-        deinit {
-            pthread_mutex_destroy(mutex)
-            mutex.deallocate()
-        }
-
-        private func lock() {
-            pthread_mutex_lock(mutex)
-        }
-
-        private func unlock() {
-            pthread_mutex_unlock(mutex)
-        }
-
-        func withLock<T>(method: (inout R) throws -> T) throws -> T {
-            defer { unlock() }
-            lock()
-            return try method(&resource)
-        }
-
-        func withLock<T>(method: (inout R) -> T) -> T {
-            defer { unlock() }
-            lock()
-            return method(&resource)
-        }
-    }
-
-}
